services/twitter-ingestor/requirements.txt
google-cloud-pubsub==2.23.0
requests==2.31.0
